+++ user_docs/en/changes.md
@@ -90,6 +90,8 @@ As the NVDA update check URL is now configurable directly within NVDA, no replac
 
 * The `braille.filter_displaySize` extension point is deprecated.
 Please use `braille.filter_displayDimensions` instead. (#17011)
+* The `gui.message.messageBox` function is deprecated.
+Use `gui.message.MessageDialog` instead.
 
 ## 2024.4